refactor: 💡 Change the location for components in UI
Change the location to a better sort with separated folders for
components made by ourselfs and components from libraries
