--- packages/ui/src/home/components/promotionalSlider/index.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-import { CosplayGallery } from './cosplay-gallery'
-
-export const PromotionalSlider = () => {
-  return (
-    <div className="flex items-center justify-center p-4">
-      <CosplayGallery />
-      {/* <CosplayGalleryWithRealImages /> */}
-    </div>
-  )
-}
