--- packages/ui/src/home/components/testimonial-card/hooks/use-media-query.ts
@@ -1,29 +1,0 @@
-"use client"
-
-import { useState, useEffect } from "react"
-
-export function useMediaQuery(query: string): boolean {
-  const [matches, setMatches] = useState(false)
-
-  useEffect(() => {
-    // Set initial value on client side
-    const media = window.matchMedia(query)
-    setMatches(media.matches)
-
-    // Create event listener for changes
-    const listener = (e: MediaQueryListEvent) => {
-      setMatches(e.matches)
-    }
-
-    // Add listener
-    media.addEventListener("change", listener)
-
-    // Clean up
-    return () => {
-      media.removeEventListener("change", listener)
-    }
-  }, [query])
-
-  return matches
-}
-
