--- packages/ui/src/home/layout/index.tsx
@@ -1,18 +1,0 @@
-import type React from "react";
-import type { ReactNode } from "react";
-import { Navbar } from "../components/navbar";
-import { FuturisticFooter } from "../components/footer/footer";
-
-interface LayoutHomeProps {
-  children: ReactNode;
-}
-
-export const LayoutHome: React.FC<LayoutHomeProps> = ({ children }) => {
-  return (
-    <>
-      <Navbar />
-      {children}
-      <FuturisticFooter />
-    </>
-  );
-};
